+++ app/models/board.rb
@@ -2,8 +2,10 @@ class Board < ApplicationRecord
   has_many :participations, as: :participant, dependent: :destroy
   has_many :lists, -> { order(position: :asc) }, dependent: :destroy
 
-  validates :title, presence: true, length: { maximum: 100 }
-  before_create :set_slug
+  validates :slug, presence: true, length: { is: 8 }, uniqueness: true
+  validates :title, presence: true, length: { maximum: 512 }
+  
+  before_validation :set_slug, on: :create
 
   def to_param
     slug