+++ spec/factories/board.rb
@@ -1,0 +1,5 @@
+FactoryBot.define do
+  factory :board do
+    title  { "a board" }
+  end
+end